FIX: Attemp to refactor using clean code arch
