+++ domain/model/vars.go
@@ -1,0 +1,8 @@
+package model
+
+// Structo to handle vars in request
+type Vars struct {
+	Items          int    `json:"items"`
+	ItemsPerWorker int    `json:"itemsPerWorker"`
+	Type           string `json:"type"`
+}
